+++ app/actions/tables.js
@@ -1,5 +1,8 @@
 // @flow
-import type { Action, TableNormalizePayload, IdString } from '../types';
+import type {
+  Action, IdString,
+  Table, TableNormalizePayload, TableDataNormalizedPayload,
+} from '../types';
 
 export const fetchTablesRequest = (payload: ?IdString): Action =>
   ({
@@ -23,3 +26,25 @@ export const setTableNameSearchKey = (payload: ?IdString): Action =>
     payload,
   });
 
+export const fetchTableData = (payload: Table): Action =>
+  ({
+    type: 'tables/FETCH_TABLE_DATA_REQUEST',
+    payload,
+  });
+
+export const selectTable = (payload: string): Action =>
+  ({
+    type: 'tables/SELECT_TABLE',
+    payload,
+  });
+
+export const setTableData = (payload: TableDataNormalizedPayload): Action =>
+  ({
+    type: 'tables/SET_TABLE_DATA',
+    payload,
+  });
+
+export const resetSelectTable = (): Action =>
+  ({
+    type: 'tables/RESET_SELECT_TABLE',
+  });
