+++ app/containers/Main/FavoritesSwitcher/FavoritesSwitcher.jsx
@@ -8,7 +8,6 @@ import * as favoritesActions from '../../../actions/favorites';
 import * as connectActions from '../../../actions/connect';
 import * as uiActions from '../../../actions/ui';
 import * as tablesActions from '../../../actions/tables';
-import * as currentTableActions from '../../../actions/currentTable';
 import type { Dispatch, Favorites, State, IdString } from '../../../types';
 import { getFavorites } from '../../../selectors/favorites';
 
@@ -91,7 +90,7 @@ class FavoritesSwitcher extends Component {
 
 function mapDispatchToProps(dispatch: Dispatch): { [key: string]: Function } {
   return bindActionCreators(
-    { ...favoritesActions, ...connectActions, ...uiActions, ...tablesActions, ...currentTableActions }, dispatch,
+    { ...favoritesActions, ...connectActions, ...uiActions, ...tablesActions }, dispatch,
   );
 }
