+++ app/sagas/tables.js
@@ -1,13 +1,20 @@
-import { take, cps, put } from 'redux-saga/effects';
+import { delay } from 'redux-saga';
+import { take, takeEvery, cps, put } from 'redux-saga/effects';
 
 import {
   fillTables as fillTablesAction,
+  selectTable as selectTableAction,
+  setTableData as setTableDataAction,
+  fetchTableData as fetchTableDataAction,
 } from '../actions/tables';
 import { executeSQL } from '../utils/pgDB';
 
 import { addFavoriteTablesQuantity } from '../actions/favorites';
-import { selectTable, fetchTableData } from '../actions/currentTable';
-import { toggleIsFetchedTables } from '../actions/ui';
+import {
+  toggleIsFetchedTables as toggleIsFetchedTablesAction,
+  setDataForMeasure as setDataForMeasureAction,
+} from '../actions/ui';
+
 
 export function* fetchTables() {
   while (true) {
@@ -37,7 +44,7 @@ export function* fetchTables() {
       map: tables,
     }));
 
-    yield put(toggleIsFetchedTables(true));
+    yield put(toggleIsFetchedTablesAction(true));
 
     if (payload) {
       yield put(addFavoriteTablesQuantity({
@@ -46,12 +53,70 @@ export function* fetchTables() {
     }
 
     if (tablesIds.length) {
-      yield put(selectTable(tables[tablesIds[0]].id));
-      yield put(fetchTableData({
+      yield put(selectTableAction(tables[tablesIds[0]].id));
+      yield put(fetchTableDataAction({
         id: tables[tablesIds[0]].id,
         tableName: tables[tablesIds[0]].tableName,
         isFetched: false,
       }));
     }
   }
+}
+
+function* fetchTableData({ payload: { id, tableName, isFetched } }) {
+  if (!isFetched) {
+    const query = `
+      SELECT *
+      FROM ${tableName}
+    `;
+    const result = yield cps(executeSQL, query, []);
+    const rows = {};
+    const dataForMeasure = {};
+
+    const fields = {};
+    const fieldsIds = result.fields.map((field, index) => {
+      const fId = index.toString();
+      fields[fId] = {
+        fieldName: field.name,
+      };
+      dataForMeasure[field.name] = {
+        value: field.name.toString(),
+        name: field.name.toString(),
+        isMeasured: false,
+        width: null,
+      };
+      return fId;
+    });
+
+    const rowsIds = result.rows.map((row, index) => {
+      const rId = index.toString();
+      rows[rId] = {
+        ...row,
+      };
+      Object.keys(row).forEach(key => {
+        const value = row[key] ? row[key].toString() : '';
+        if (dataForMeasure[key].value.length < value.length) {
+          dataForMeasure[key].value = row[key].toString();
+          dataForMeasure[key].isMeasured = false;
+        }
+      });
+      return rId;
+    });
+
+    yield put(setDataForMeasureAction(dataForMeasure));
+    yield delay(100); // This delay needs to measure cells
+
+    yield put(setTableDataAction({
+      id,
+      rowsIds,
+      rows,
+      fieldsIds,
+      fields,
+      isFetched: true,
+    }));
+  }
+}
+
+export function* fetchTableDataWatch() {
+  yield takeEvery('tables/FETCH_TABLE_DATA_REQUEST', fetchTableData);
 }
